Rename unicode_align crate to unicode_align_test
