+++ src/redisearch_rs/unicode_align_test/src/lib.rs
@@ -635,7 +635,7 @@ impl Report {
     /// Render the aggregate report as human-readable text.
     pub fn render(&self) -> String {
         let mut out = String::new();
-        let _ = writeln!(out, "=== unicode_align report ===");
+        let _ = writeln!(out, "=== unicode_align_test report ===");
         let pct = if self.total > 0 {
             100.0 * self.diverged as f64 / self.total as f64
         } else {
