+++ src/redisearch_rs/unicode_align_test/tests/codepoint_sweep.rs
@@ -14,7 +14,7 @@
 //! like final sigma don't interfere) and let the corpus aggregator emit the
 //! full codepoint divergence table.
 
-use unicode_align::run_corpus;
+use unicode_align_test::run_corpus;
 
 #[test]
 fn sweep_all_codepoints() {
