+++ src/redisearch_rs/unicode_align_test/tests/single_strings_lower.rs
@@ -7,21 +7,30 @@
  * GNU Affero General Public License v3 (AGPLv3).
 */
 
-//! Hand-picked strings that exercise well-known libnu vs ICU divergences.
+//! Hand-picked strings exercising known libnu vs ICU lowercase divergences.
 //!
-//! These tests print their results; they don't gate CI on divergence
-//! (the divergence is the point — this is a comparison harness, not a
-//! regression check).
+//! Sister to `single_strings` (fold). The shared cases (sharp s, sigma,
+//! Turkish I, ligatures) test very different surface area under lowercase:
+//! folding canonicalises Σ/σ/ς to a single codepoint regardless of position,
+//! whereas lowercase preserves uppercase-vs-lowercase distinctions and is
+//! context-sensitive for final sigma (in ICU) — context that libnu's
+//! unconditional `nu_tolower` ignores.
 
-use unicode_align::{compare, run_corpus};
+use unicode_align_test::{compare_lower, run_corpus_lower};
 
 fn known_cases() -> Vec<(&'static str, &'static str)> {
     vec![
         ("sharp s (German)", "Straße"),
         ("uppercase sharp s", "STRAẞE"),
-        ("Greek final sigma", "ΟΔΥΣΣΕΥΣ"),
+        (
+            "Greek final sigma (word-final triggers ς in ICU)",
+            "ΟΔΥΣΣΕΥΣ",
+        ),
         ("Greek mixed", "Ὀδυσσεύς"),
-        ("Turkish dotted I", "İstanbul"),
+        (
+            "Turkish dotted I (root locale lowercases to i\u{0307})",
+            "İstanbul",
+        ),
         ("Turkish dotless i (uppercase)", "IZMIR"),
         ("Latin ffi ligature", "ﬃ"),
         ("Latin ff ligature", "ﬀ"),
@@ -37,21 +46,21 @@ fn known_cases() -> Vec<(&'static str, &'static str)> {
 }
 
 #[test]
-fn print_known_divergences() {
-    println!("\n=== single-string known divergence cases ===");
+fn print_known_lowercase_divergences() {
+    println!("\n=== single-string lowercase divergence cases ===");
     for (name, input) in known_cases() {
-        let diff = compare(input);
+        let diff = compare_lower(input);
         println!("\n## {name}");
         print!("{}", diff.render());
     }
 }
 
 #[test]
-fn aggregate_known_cases() {
+fn aggregate_known_lowercase_cases() {
     let inputs: Vec<String> = known_cases()
         .into_iter()
         .map(|(_, s)| s.to_owned())
         .collect();
-    let report = run_corpus(inputs);
+    let report = run_corpus_lower(inputs);
     print!("{}", report.render());
 }
